feat: added two separate pages for homepage and details page with routing
