+++ 03-miscellaneous/react-async-ui/tsconfig.app.json
@@ -1,6 +1,5 @@
 {
     "compilerOptions": {
-        "ignoreDeprecations": "6.0",
         "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.app.tsbuildinfo",
         "target": "es2023",
         "lib": ["ES2023", "DOM", "DOM.Iterable"],
@@ -21,7 +20,6 @@
         "noUnusedParameters": true,
         "erasableSyntaxOnly": true,
         "noFallthroughCasesInSwitch": true,
-        "baseUrl": ".",
         "paths": {
             "@/*": ["./src/*"]
         }
